split ui components in c and header file,t o reduce compilation times of frequently used ui components
