+++ src/ui/components/meson.build
@@ -1,8 +1,14 @@
 src_files += files(
+    'button.cpp',
     'button.hpp',
+    'image_view.cpp',
     'image_view.hpp',
+    'label.cpp',
     'label.hpp',
+    'link_label.cpp',
     'link_label.hpp',
+    'slider.cpp',
     'slider.hpp',
+    'textinput.cpp',
     'textinput.hpp',
 )
