+++ src/ui/layouts/meson.build
@@ -1,1 +1,10 @@
-src_files += files('focus_layout.hpp', 'grid_layout.hpp', 'scroll_layout.hpp', 'tile_layout.hpp')
+src_files += files(
+    'focus_layout.cpp',
+    'focus_layout.hpp',
+    'grid_layout.cpp',
+    'grid_layout.hpp',
+    'scroll_layout.cpp',
+    'scroll_layout.hpp',
+    'tile_layout.cpp',
+    'tile_layout.hpp',
+)
